Add ability to set number of lines per page

+++ include/display.h
@@ -6,8 +6,8 @@
 void print_status(const struct lotsctl *const);
 void clear_status(void);
 
-void move_forwards(struct lotsctl *const, int);
-void move_backwards(struct lotsctl *const, int);
+void move_forwards(struct lotsctl *const, unsigned long);
+void move_backwards(struct lotsctl *const, unsigned long);
 
 void display_file(struct lotsctl *const, const char *const);
